Optimize git meta root and forward command.
Both the forward command and root command currently load NodeGit
for no specific reason. Removing this loading and breaking out
non-NodeGit utility methods makes these commands significantly faster.
